requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements.in/base.in
#
appdirs==1.4.4
    # via
    #   -r requirements.in/base.in
    #   virtualenv
asgiref==3.3.2
    # via
    #   -r requirements.in/base.in
    #   django
attrs==20.3.0
    # via
    #   -r requirements.in/base.in
    #   pytest
certifi==2020.12.5
    # via
    #   -r requirements.in/base.in
    #   requests
    #   sentry-sdk
cfgv==3.2.0
    # via
    #   -r requirements.in/base.in
    #   pre-commit
chardet==4.0.0
    # via
    #   -r requirements.in/base.in
    #   requests
click==7.1.2
    # via
    #   -r requirements.in/base.in
    #   pip-tools
configargparse==1.4.1
    # via -r requirements.in/base.in
convertdate==2.3.2
    # via
    #   -r requirements.in/base.in
    #   holidays
coverage[toml]==5.5
    # via
    #   -r requirements.in/base.in
    #   pytest-cov
defusedxml==0.6.0
    # via
    #   -r requirements.in/base.in
    #   odfpy
diff-match-patch==20200713
    # via
    #   -r requirements.in/base.in
    #   django-import-export
distlib==0.3.1
    # via
    #   -r requirements.in/base.in
    #   virtualenv
django==3.2.5
    # via
    #   -r requirements.in/base.in
    #   django-import-export
    #   django-reversion
    #   django-staff-sso-client
    #   djangorestframework
django-environ==0.4.5
    # via -r requirements.in/base.in
django-import-export==2.5.0
    # via -r requirements.in/base.in
django-ipware==3.0.2
    # via
    #   -r requirements.in/base.in
    #   django-log-formatter-ecs
django-log-formatter-ecs==0.0.5
    # via -r requirements.in/base.in
django-reversion==3.0.9
    # via -r requirements.in/base.in
django-simple-history==3.0.0
    # via -r requirements.in/base.in
django-staff-sso-client==3.1.0
    # via -r requirements.in/base.in
django-webpack-loader==1.0.0
    # via -r requirements.in/base.in
djangorestframework==3.12.2
    # via
    #   -r requirements.in/base.in
    #   hawkrest
et-xmlfile==1.0.1
    # via
    #   -r requirements.in/base.in
    #   openpyxl
factory-boy==3.2.0
    # via -r requirements.in/base.in
faker==6.0.0
    # via
    #   -r requirements.in/base.in
    #   factory-boy
filelock==3.0.12
    # via
    #   -r requirements.in/base.in
    #   virtualenv
hawkrest==1.0.1
    # via -r requirements.in/base.in
hijri-converter==2.1.1
    # via
    #   -r requirements.in/base.in
    #   holidays
holidays==0.10.5.2
    # via -r requirements.in/base.in
identify==1.5.13
    # via
    #   -r requirements.in/base.in
    #   pre-commit
idna==2.10
    # via
    #   -r requirements.in/base.in
    #   requests
importlib-metadata==4.0.1
    # via -r requirements.in/base.in
iniconfig==1.1.1
    # via
    #   -r requirements.in/base.in
    #   pytest
itsdangerous==2.0.1
    # via -r requirements.in/base.in
jdcal==1.4.1
    # via
    #   -r requirements.in/base.in
    #   openpyxl
jinja2==3.0.1
    # via -r requirements.in/base.in
korean-lunar-calendar==0.2.1
    # via
    #   -r requirements.in/base.in
    #   holidays
kubi-ecs-logger==0.1.0
    # via
    #   -r requirements.in/base.in
    #   django-log-formatter-ecs
markuppy==1.14
    # via
    #   -r requirements.in/base.in
    #   tablib
markupsafe==2.0.1
    # via
    #   -r requirements.in/base.in
    #   jinja2
marshmallow==3.6.1
    # via
    #   -r requirements.in/base.in
    #   kubi-ecs-logger
mohawk==1.1.0
    # via
    #   -r requirements.in/base.in
    #   hawkrest
msgpack==1.0.2
    # via -r requirements.in/base.in
mypy-extensions==0.4.3
    # via -r requirements.in/base.in
nodeenv==1.5.0
    # via
    #   -r requirements.in/base.in
    #   pre-commit
oauthlib==3.1.0
    # via
    #   -r requirements.in/base.in
    #   requests-oauthlib
odfpy==1.4.1
    # via
    #   -r requirements.in/base.in
    #   tablib
openpyxl==3.0.6
    # via
    #   -r requirements.in/base.in
    #   tablib
packaging==20.9
    # via
    #   -r requirements.in/base.in
    #   pytest
pathspec==0.8.1
    # via -r requirements.in/base.in
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in/base.in
pluggy==0.13.1
    # via
    #   -r requirements.in/base.in
    #   pytest
pre-commit==2.10.0
    # via -r requirements.in/base.in
psutil==5.8.0
    # via -r requirements.in/base.in
psycopg2-binary==2.8.6
    # via -r requirements.in/base.in
py==1.10.0
    # via
    #   -r requirements.in/base.in
    #   pytest
pymeeus==0.5.11
    # via
    #   -r requirements.in/base.in
    #   convertdate
pyparsing==2.4.7
    # via
    #   -r requirements.in/base.in
    #   packaging
pytest==6.2.2
    # via
    #   -r requirements.in/base.in
    #   pytest-cov
    #   pytest-django
pytest-cov==2.12.0
    # via -r requirements.in/base.in
pytest-django==4.1.0
    # via -r requirements.in/base.in
python-dateutil==2.8.1
    # via
    #   -r requirements.in/base.in
    #   faker
    #   holidays
pytz==2020.5
    # via
    #   -r requirements.in/base.in
    #   convertdate
    #   django
pyyaml==5.4.1
    # via
    #   -r requirements.in/base.in
    #   pre-commit
    #   tablib
pyzmq==22.1.0
    # via -r requirements.in/base.in
regex==2020.11.13
    # via -r requirements.in/base.in
requests==2.25.1
    # via
    #   -r requirements.in/base.in
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via
    #   -r requirements.in/base.in
    #   django-staff-sso-client
sentry-sdk==0.20.3
    # via -r requirements.in/base.in
six==1.15.0
    # via
    #   -r requirements.in/base.in
    #   holidays
    #   mohawk
    #   python-dateutil
    #   virtualenv
sqlparse==0.4.2
    # via
    #   -r requirements.in/base.in
    #   django
tablib[html,ods,xls,xlsx,yaml]==3.0.0
    # via
    #   -r requirements.in/base.in
    #   django-import-export
text-unidecode==1.3
    # via
    #   -r requirements.in/base.in
    #   faker
toml==0.10.2
    # via
    #   -r requirements.in/base.in
    #   coverage
    #   pre-commit
    #   pytest
tomli==1.2.2
    # via pep517
typed-ast==1.4.2
    # via -r requirements.in/base.in
typing-extensions==3.7.4.3
    # via -r requirements.in/base.in
urllib3==1.26.5
    # via
    #   -r requirements.in/base.in
    #   requests
    #   sentry-sdk
virtualenv==20.4.0
    # via
    #   -r requirements.in/base.in
    #   pre-commit
werkzeug==2.0.1
    # via -r requirements.in/base.in
wheel==0.37.0
    # via pip-tools
whitenoise==5.2.0
    # via -r requirements.in/base.in
xlrd==2.0.1
    # via
    #   -r requirements.in/base.in
    #   tablib
xlwt==1.3.0
    # via
    #   -r requirements.in/base.in
    #   tablib
zipp==3.4.1
    # via
    #   -r requirements.in/base.in
    #   importlib-metadata
zope.event==4.5.0
    # via -r requirements.in/base.in
zope.interface==5.4.0
    # via -r requirements.in/base.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
